+++ src/engine/compression/ArithmeticCompression.java
@@ -65,7 +65,7 @@ public class ArithmeticCompression extends AbstractCompressor {
 		}
 		finally {
 			try {
-				compressionFactor = 1 - ( outputSize / fragmenter.getFileSize() );
+				compressionFactor = 1 - ( (float)outputSize / (float)fragmenter.getFileSize() );
 				outputStream.close();
 				fragmenter.close();
 			} catch (Exception e) {}
